Refine changes on how allowed networks are handled
First off, this changes the role to allow traffic to multicast address
ranges for ipv4 and ipv6.  This is critical for ipv6.  Handling ipv4 and
ipv6 the same here allows for consistency.  In addition, ipv6 link local
addresses are also allowed via the ALLOWNET chain.  Update README.md and
add some variables to assist in this process.